refactor(web): compose feed and thread-surface reads as explicit variants
Replace the ThreadReadCard variant prop with shared compound pieces so each surface declares its chrome instead of branching inside one card.
