+++ apps/web/src/components/signals/action-row/index.ts
@@ -1,6 +1,6 @@
 export { ActionRow, ActionRowSkeleton } from "./action-row";
-export type { ThreadWithRelations } from "./thread-read-card";
-export { ThreadReadCard } from "./thread-read-card";
+export type { ThreadWithAgentRead, ThreadWithRelations } from "./thread-read";
+export { FeedThreadRead, ThreadRead, ThreadSurfaceRead } from "./thread-read";
 export type { ActorContext } from "./handlers";
 export {
   acceptInlineSuggestion,
