+++ apps/web/src/components/signals/action-row/signal-reply-draft-editor.tsx
@@ -4,12 +4,26 @@ import { stringify } from "@workspace/utils/tiptap-md";
 import { PenLineIcon } from "lucide-react";
 import { AnimatePresence, motion } from "motion/react";
 import { useCallback } from "react";
+import type { ReactNode } from "react";
 
 interface Props {
   open: boolean;
   draft: string;
   contentKey: string;
   onDraftChange: (draft: string) => void;
+}
+
+/** Presentational chrome around a reply draft — live editor or static mock. */
+export function ReplyDraftChrome({ children }: { children: ReactNode }) {
+  return (
+    <div className="space-y-2 px-3 pt-2 pb-0">
+      <div className="flex items-center gap-1.5 text-xs text-foreground-secondary mt-1">
+        <PenLineIcon className="size-3.5 shrink-0" />
+        <span>Reply draft</span>
+      </div>
+      {children}
+    </div>
+  );
 }
 
 export function SignalReplyDraftEditor({
@@ -35,11 +49,7 @@ export function SignalReplyDraftEditor({
           transition={{ duration: 0.15, ease: "easeInOut" }}
           className="overflow-hidden bg-background-tertiary"
         >
-          <div className="space-y-2 px-3 pt-2 pb-0">
-            <div className="flex items-center gap-1.5 text-xs text-foreground-secondary mt-1">
-              <PenLineIcon className="size-3.5 shrink-0" />
-              <span>Reply draft</span>
-            </div>
+          <ReplyDraftChrome>
             <div className="max-h-52 min-h-0 overflow-y-auto text-sm">
               <EditableRichText
                 key={contentKey}
@@ -48,7 +58,7 @@ export function SignalReplyDraftEditor({
                 className="[&_.ProseMirror]:p-0 [&_.ProseMirror]:min-h-0"
               />
             </div>
-          </div>
+          </ReplyDraftChrome>
         </motion.div>
       ) : null}
     </AnimatePresence>
