--- apps/web/src/components/signals/action-row/thread-read-card.tsx
@@ -1,860 +1,0 @@
-import type { InferLiveObject } from "@live-state/sync";
-import { useLiveQuery } from "@live-state/sync/client";
-import { Link } from "@tanstack/react-router";
-import {
-  ACTION_KIND_LABEL,
-  ACTION_KIND_VERB,
-  fingerprintAgentRead,
-  sanitizeAgentReadReasoning,
-  urgencyTierFromScore,
-} from "@workspace/schemas/signals";
-import type {
-  Action,
-  InlineSuggestion,
-  ReplyAction,
-  ThreadRead,
-} from "@workspace/schemas/signals";
-import { Avatar } from "@workspace/ui/components/avatar";
-import { ActionButton } from "@workspace/ui/components/button";
-import { Checkbox } from "@workspace/ui/components/checkbox";
-import {
-  HoverCard,
-  HoverCardContent,
-  HoverCardTrigger,
-} from "@workspace/ui/components/hover-card";
-import {
-  treeContentClassName,
-  TreeJoin,
-  treeRowClassName,
-  TreeSkip,
-  TREE_ROW_GAP_PX,
-} from "@workspace/ui/components/tree";
-import { cn, formatRelativeTime } from "@workspace/ui/lib/utils";
-import type { schema } from "api/schema";
-import { Brain, Check, X } from "lucide-react";
-import { useEffect, useId, useMemo, useState } from "react";
-import { toast } from "sonner";
-
-import { ThreadSummaryHoverCard } from "~/components/chips";
-import { RichMarkdown } from "~/components/markdown/rich-markdown";
-import { query } from "~/lib/live-state";
-import { buildThreadParam } from "~/utils/thread";
-
-import { ActionRow } from "./action-row";
-import {
-  acceptInlineSuggestion,
-  acceptThreadRead,
-  dismissInlineSuggestion,
-  dismissThreadRead,
-} from "./handlers";
-import type { ActorContext } from "./handlers";
-import { SignalReplyDraftEditor } from "./signal-reply-draft-editor";
-
-export type ThreadWithRelations = InferLiveObject<
-  typeof schema.thread,
-  {
-    author: { include: { user: true } };
-    assignedUser: { include: { user: true } };
-  }
->;
-
-interface ThreadReadCardBase {
-  thread: ThreadWithRelations & { agentRead: ThreadRead };
-  ctx: ActorContext;
-}
-
-type Props =
-  | (ThreadReadCardBase & {
-      /** Includes thread identity chrome and a dismiss X. */
-      variant?: "feed";
-      onClose?: never;
-    })
-  | (ThreadReadCardBase & {
-      /** Drops thread chrome (already on the thread). X closes the panel. */
-      variant: "thread";
-      onClose: () => void;
-    });
-
-function ThreadRef({ thread }: { thread: ThreadWithRelations }) {
-  return (
-    <ThreadSummaryHoverCard thread={thread}>
-      <Link
-        to="/app/threads/$id"
-        params={{ id: buildThreadParam(thread) }}
-        className="inline-flex w-fit items-center gap-1.5 text-sm"
-      >
-        <Avatar
-          variant="user"
-          size="md"
-          fallback={thread.author?.name}
-          src={thread.author?.user?.image ?? undefined}
-        />
-        <span className="text-foreground-primary">{thread.name}</span>
-        {thread.shortId !== null && (
-          <span className="text-foreground-secondary tabular-nums">
-            #{thread.shortId}
-          </span>
-        )}
-      </Link>
-    </ThreadSummaryHoverCard>
-  );
-}
-
-interface SelectedAction {
-  action: Action;
-  index: number;
-}
-
-function primaryReplyAction(primary: Action[]): ReplyAction | undefined {
-  return primary.find(
-    (action): action is ReplyAction => action.kind === "reply"
-  );
-}
-
-function primaryReplyDraftMarkdown(primary: Action[]): string {
-  return primaryReplyAction(primary)?.draftMarkdown ?? "";
-}
-
-function capitalize(value: string): string {
-  return value.charAt(0).toUpperCase() + value.slice(1);
-}
-
-/**
- * Selected bundle actions in display order: reply always leads, the rest keep
- * their original bundle order.
- */
-function orderReplyFirst(
-  primary: Action[],
-  selected: ReadonlySet<number>
-): SelectedAction[] {
-  return primary
-    .map((action, index): SelectedAction => ({ action, index }))
-    .filter(({ index }) => selected.has(index))
-    .toSorted((a, b) => {
-      const aReply = a.action.kind === "reply" ? 0 : 1;
-      const bReply = b.action.kind === "reply" ? 0 : 1;
-      if (aReply !== bReply) {
-        return aReply - bReply;
-      }
-      return a.index - b.index;
-    });
-}
-
-/**
- * Compose the compound-action button copy from the current selection, e.g.
- * "Reply and close", "Reply and do 2 actions", or — once the reply editor is
- * open — "Send and close". The leading verb becomes "Send" while editing.
- */
-function compoundButtonLabel(
-  ordered: SelectedAction[],
-  replyEditorOpen: boolean
-): string {
-  if (ordered.length === 0) {
-    return "Select an action";
-  }
-
-  const verbAt = (entry: SelectedAction, position: number): string => {
-    if (entry.action.kind === "reply" && replyEditorOpen) {
-      return position === 0 ? "Send" : "send";
-    }
-    return ACTION_KIND_VERB[entry.action.kind];
-  };
-
-  const first = capitalize(verbAt(ordered[0], 0));
-  if (ordered.length === 1) {
-    return first;
-  }
-  if (ordered.length === 2) {
-    return `${first} and ${verbAt(ordered[1], 1)}`;
-  }
-  return `${first} and do ${ordered.length - 1} actions`;
-}
-
-interface CompoundActionButtonProps {
-  label: string;
-  /** Hard-disabled (e.g. busy executing): blocks the trigger entirely. */
-  disabled: boolean;
-  /**
-   * Execution is invalid right now (nothing selected, or empty reply). The
-   * click is gated, but for a compound bundle the selector stays reachable so
-   * the user can re-select an action and recover.
-   */
-  executeDisabled: boolean;
-  onClick: () => void;
-  /** The full bundle to offer for selection, or null for a single action. */
-  actions: Action[] | null;
-  selectedIndices: ReadonlySet<number>;
-  lockedIndex: number;
-  disableToggles: boolean;
-  onToggle: (index: number) => void;
-}
-
-function CompoundActionButton({
-  label,
-  disabled,
-  executeDisabled,
-  onClick,
-  actions,
-  selectedIndices,
-  lockedIndex,
-  disableToggles,
-  onToggle,
-}: CompoundActionButtonProps) {
-  const fieldId = useId();
-
-  if (!actions) {
-    return (
-      <ActionButton
-        size="sm"
-        variant="primary"
-        onClick={onClick}
-        disabled={disabled || executeDisabled}
-      >
-        {label}
-      </ActionButton>
-    );
-  }
-
-  return (
-    <HoverCard>
-      <HoverCardTrigger
-        render={
-          <ActionButton
-            size="sm"
-            variant="primary"
-            onClick={executeDisabled ? undefined : onClick}
-            disabled={disabled}
-            aria-disabled={executeDisabled || undefined}
-            className={cn(executeDisabled && "opacity-60")}
-          />
-        }
-      >
-        {label}
-      </HoverCardTrigger>
-      <HoverCardContent
-        align="end"
-        className="flex w-60 flex-col gap-0.5 p-1.5"
-      >
-        <p className="px-1.5 py-1 text-xs text-foreground-secondary">
-          Actions in this bundle
-        </p>
-        {actions.map((action, index) => {
-          const locked = index === lockedIndex;
-          const checkboxId = `${fieldId}-${index}`;
-          return (
-            <div
-              key={`${action.kind}:${index}`}
-              className="flex items-center gap-2 rounded-sm px-1.5 py-1 hover:bg-accent has-disabled:opacity-60"
-            >
-              <Checkbox
-                id={checkboxId}
-                checked={selectedIndices.has(index)}
-                disabled={locked || disableToggles}
-                onCheckedChange={() => onToggle(index)}
-              />
-              <label
-                htmlFor={checkboxId}
-                className="flex-1 cursor-pointer select-none text-sm text-foreground-primary has-disabled:cursor-default"
-              >
-                {ACTION_KIND_LABEL[action.kind]}
-              </label>
-            </div>
-          );
-        })}
-      </HoverCardContent>
-    </HoverCard>
-  );
-}
-
-interface ResolvedInlineSuggestion {
-  suggestion: InlineSuggestion;
-  name: string;
-  color: string;
-}
-
-/**
- * Label suggestions rendered as dashed chips,
- * mirroring the thread toolbar's quick-actions style: click a chip to apply,
- * hover for detail, and reveal apply-all / ignore-all on row hover.
- */
-function InlineSuggestionsRow({
-  suggestions,
-  organizationId,
-  busy,
-  onAccept,
-  onDismiss,
-}: {
-  suggestions: InlineSuggestion[];
-  organizationId: string;
-  busy: boolean;
-  onAccept: (suggestion: InlineSuggestion) => void | Promise<void>;
-  onDismiss: (suggestion: InlineSuggestion) => void | Promise<void>;
-}) {
-  const labels = useLiveQuery(query.label.where({ organizationId }));
-
-  const resolved = useMemo<ResolvedInlineSuggestion[]>(() => {
-    const labelById = new Map((labels ?? []).map((label) => [label.id, label]));
-    const result: ResolvedInlineSuggestion[] = [];
-    for (const suggestion of suggestions) {
-      const label = labelById.get(suggestion.action.labelId);
-      if (!label) {
-        // Keep the suggestion dismissible even when its label was deleted,
-        // disabled, or hasn't loaded yet — otherwise it becomes a stuck row.
-        result.push({
-          color: "var(--muted-foreground)",
-          name: "Unknown label",
-          suggestion,
-        });
-        continue;
-      }
-      result.push({
-        color: label.color,
-        name: label.name,
-        suggestion,
-      });
-    }
-    return result;
-  }, [labels, suggestions]);
-
-  if (resolved.length === 0) {
-    return null;
-  }
-
-  const chipContent = (item: ResolvedInlineSuggestion) => (
-    <>
-      <div
-        className="size-2 rounded-full"
-        style={{ backgroundColor: item.color }}
-      />
-      {item.name}
-    </>
-  );
-
-  return (
-    <div className="group flex min-w-0 flex-wrap items-center gap-x-4 gap-y-2 text-sm">
-      <span className="shrink-0 text-foreground-secondary">
-        Quick suggestions
-      </span>
-      <div className="flex min-w-0 flex-wrap items-center gap-2">
-        {resolved.map((item) => (
-          <HoverCard key={item.suggestion.id}>
-            <HoverCardTrigger
-              render={
-                <ActionButton
-                  variant="ghost"
-                  size="sm"
-                  className="border border-dashed border-input dark:hover:bg-foreground-tertiary/15"
-                  onClick={() => onAccept(item.suggestion)}
-                  disabled={busy}
-                />
-              }
-            >
-              {chipContent(item)}
-            </HoverCardTrigger>
-            <HoverCardContent className="min-w-72 w-full max-w-96 flex flex-col gap-3">
-              <div className="text-xs flex flex-col gap-1">
-                <div className="text-foreground-secondary">Suggested label</div>
-                <div className="border border-dashed flex items-center w-fit h-6 rounded-sm gap-1.5 px-2 has-[>svg:first-child]:pl-1.5 text-xs bg-foreground-tertiary/15">
-                  {chipContent(item)}
-                </div>
-              </div>
-              <ActionButton
-                variant="outline"
-                size="sm"
-                className="w-full"
-                onClick={() => onAccept(item.suggestion)}
-                disabled={busy}
-              >
-                Apply suggestion
-              </ActionButton>
-            </HoverCardContent>
-          </HoverCard>
-        ))}
-        <div className="flex items-center gap-0 opacity-0 transition-opacity pointer-events-none group-hover:opacity-100 group-hover:pointer-events-auto focus-within:opacity-100 focus-within:pointer-events-auto">
-          <ActionButton
-            variant="ghost"
-            size="icon-sm"
-            tooltip="Apply all"
-            className="text-foreground-secondary"
-            onClick={async () => {
-              // Run sequentially so a single busy state covers the whole batch
-              // instead of concurrent requests racing to clear it.
-              for (const item of resolved) {
-                await onAccept(item.suggestion);
-              }
-            }}
-            disabled={busy}
-          >
-            <Check />
-          </ActionButton>
-          <ActionButton
-            variant="ghost"
-            size="icon-sm"
-            tooltip="Ignore all"
-            className="text-foreground-secondary"
-            onClick={async () => {
-              for (const item of resolved) {
-                await onDismiss(item.suggestion);
-              }
-            }}
-            disabled={busy}
-          >
-            <X />
-          </ActionButton>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-function AgentReadReasoningTrigger({ reasoning }: { reasoning: string }) {
-  const trimmed = sanitizeAgentReadReasoning(reasoning);
-  if (!trimmed) {
-    return null;
-  }
-
-  return (
-    <HoverCard>
-      <HoverCardTrigger
-        render={
-          <ActionButton
-            size="sm"
-            variant="ghost"
-            tooltip="Agent reasoning"
-            aria-label="View agent reasoning"
-          />
-        }
-      >
-        <Brain className="size-3.5" />
-      </HoverCardTrigger>
-      <HoverCardContent className="max-w-96 w-full flex flex-col gap-2">
-        <p className="text-xs font-medium text-foreground-primary">Reasoning</p>
-        <RichMarkdown
-          content={trimmed}
-          preset="inline"
-          className="text-xs text-foreground-secondary"
-        />
-      </HoverCardContent>
-    </HoverCard>
-  );
-}
-
-function formatErrorMessage(error: unknown): string {
-  const message = error instanceof Error ? error.message : "";
-  if (message.includes("STALE_AGENT_READ")) {
-    return "This signal changed in the background. Refresh and try again.";
-  }
-  // A read can outlive the configuration it was built against: the org may have
-  // cleared its default issue target (or disabled the tracker) after synthesis
-  // proposed create_issue. Retrying will never work, so say what to change
-  // instead of offering the generic try-again.
-  if (message.includes("DEFAULT_ISSUE_TARGET_NOT_CONFIGURED")) {
-    return "No issue target is available. Connect a repository in Integrations settings.";
-  }
-  if (message.includes("ISSUE_TRACKER_NOT_CONFIGURED")) {
-    return "This organization has no configured issue tracker.";
-  }
-  // Raised by the connector, which owns target validation — core treats the
-  // integration config as opaque and cannot check the target up front.
-  if (message.includes("REPOSITORY_NOT_CONNECTED")) {
-    return "The configured issue target is no longer connected. Pick another in Integrations settings.";
-  }
-  return "Could not apply this signal. Please try again.";
-}
-
-export function ThreadReadCard({
-  thread,
-  ctx,
-  variant = "feed",
-  onClose,
-}: Props) {
-  const [busyKey, setBusyKey] = useState<string | null>(null);
-  // Which reply is being previewed/edited: the primary bundle's reply, or a
-  // specific alternative reply. Both open the same inline draft editor.
-  const [replyTarget, setReplyTarget] = useState<
-    { kind: "primary" } | { kind: "alternative"; index: number } | null
-  >(null);
-  const replyEditorOpen = replyTarget !== null;
-  const editingPrimaryReply = replyTarget?.kind === "primary";
-  const editingAlternativeReply = replyTarget?.kind === "alternative";
-  const read = thread.agentRead;
-  const readFingerprint = fingerprintAgentRead(read);
-  const replyIndex = read.primary.findIndex(
-    (action) => action.kind === "reply"
-  );
-  const isCompound = read.primary.length > 1;
-  const [selectedIndices, setSelectedIndices] = useState<Set<number>>(
-    () => new Set(read.primary.map((_, index) => index))
-  );
-  const [replyDraft, setReplyDraft] = useState(() =>
-    primaryReplyDraftMarkdown(read.primary)
-  );
-  const [replyEditorRevision, setReplyEditorRevision] = useState(0);
-  const inlineSuggestions = (thread.inlineSuggestions ?? []).filter(
-    (suggestion) =>
-      !suggestion.dismissedAt &&
-      // `apply_label` is the only kind the inline track writes (ADR 0014), and
-      // the row below reads `action.labelId`. Rows persisted by the retired
-      // status track would otherwise render as an "Unknown label" chip.
-      suggestion.action.kind === "apply_label"
-  );
-
-  useEffect(() => {
-    setReplyTarget(null);
-    setSelectedIndices(new Set(read.primary.map((_, index) => index)));
-    setReplyDraft(primaryReplyDraftMarkdown(read.primary));
-    setReplyEditorRevision((revision) => revision + 1);
-    // Fingerprint already covers primary/alternatives content. Depending on
-    // `read.primary` resets every card on each live-query identity change.
-  }, [readFingerprint]);
-
-  const orderedSelected = useMemo(
-    () => orderReplyFirst(read.primary, selectedIndices),
-    [read.primary, selectedIndices]
-  );
-  const selectionIncludesReply =
-    replyIndex !== -1 && selectedIndices.has(replyIndex);
-
-  // When the primary bundle is a lone reply, a reply-only alternative is
-  // redundant — it offers the same "just reply" as the primary — so drop it.
-  // Indices are kept intact so accept/edit handlers still map to read.alternatives.
-  const primaryIsJustReply =
-    read.primary.length === 1 && read.primary[0]?.kind === "reply";
-  const visibleAlternatives = useMemo(
-    () =>
-      (read.alternatives ?? [])
-        .map((alternative, index) => ({ alternative, index }))
-        .filter(
-          ({ alternative }) =>
-            !(primaryIsJustReply && alternative.kind === "reply")
-        ),
-    [read.alternatives, primaryIsJustReply]
-  );
-
-  const handleAcceptSelected = async (replyDraftValue?: string) => {
-    const indices = [...selectedIndices].toSorted((a, b) => a - b);
-    if (indices.length === 0) {
-      return;
-    }
-    setBusyKey("primary");
-    try {
-      await acceptThreadRead({
-        ctx,
-        read,
-        replyDraft: replyDraftValue,
-        selection: { primaryActionIndices: indices },
-        threadId: thread.id,
-      });
-      setReplyTarget(null);
-    } catch (error) {
-      toast.error(formatErrorMessage(error));
-    } finally {
-      setBusyKey(null);
-    }
-  };
-
-  const handlePrimaryClick = () => {
-    if (selectionIncludesReply) {
-      // Reseed the shared draft so an earlier alternative-reply flow can't leak
-      // stale text into the primary editor.
-      setReplyDraft(primaryReplyDraftMarkdown(read.primary));
-      setReplyEditorRevision((revision) => revision + 1);
-      setReplyTarget({ kind: "primary" });
-      return;
-    }
-    void handleAcceptSelected();
-  };
-
-  const handleCancelReplyEditor = () => {
-    setReplyTarget(null);
-    setReplyDraft(primaryReplyDraftMarkdown(read.primary));
-    setReplyEditorRevision((revision) => revision + 1);
-  };
-
-  const handleSendReply = () => {
-    const trimmed = replyDraft.trim();
-    if (trimmed.length === 0) {
-      return;
-    }
-    if (replyTarget?.kind === "alternative") {
-      void handleAcceptAlternative(replyTarget.index, trimmed);
-      return;
-    }
-    void handleAcceptSelected(trimmed);
-  };
-
-  const toggleAction = (index: number) => {
-    // The primary reply stays locked-in while its editor is expanded.
-    if (editingPrimaryReply && index === replyIndex) {
-      return;
-    }
-    setSelectedIndices((prev) => {
-      const next = new Set(prev);
-      if (next.has(index)) {
-        next.delete(index);
-      } else {
-        next.add(index);
-      }
-      return next;
-    });
-  };
-
-  const handleAcceptAlternative = async (
-    alternativeIndex: number,
-    replyDraftValue?: string
-  ) => {
-    setBusyKey(`alt:${alternativeIndex}`);
-    try {
-      await acceptThreadRead({
-        ctx,
-        read,
-        replyDraft: replyDraftValue,
-        selection: { alternativeIndex },
-        threadId: thread.id,
-      });
-      setReplyTarget(null);
-    } catch (error) {
-      toast.error(formatErrorMessage(error));
-    } finally {
-      setBusyKey(null);
-    }
-  };
-
-  // A reply alternative opens the same inline preview editor as the primary
-  // reply (seeded with its own draft); any other alternative applies directly.
-  const handleAlternativeClick = (alternativeIndex: number) => {
-    const alternative = read.alternatives?.[alternativeIndex];
-    if (alternative?.kind === "reply") {
-      setReplyTarget({ index: alternativeIndex, kind: "alternative" });
-      setReplyDraft(alternative.draftMarkdown ?? "");
-      setReplyEditorRevision((revision) => revision + 1);
-      return;
-    }
-    void handleAcceptAlternative(alternativeIndex);
-  };
-
-  const handleDismissRead = async () => {
-    setBusyKey("dismiss");
-    try {
-      await dismissThreadRead({
-        ctx,
-        read,
-        threadId: thread.id,
-      });
-    } catch (error) {
-      toast.error(formatErrorMessage(error));
-    } finally {
-      setBusyKey(null);
-    }
-  };
-
-  const handleInlineAccept = async (suggestion: InlineSuggestion) => {
-    setBusyKey(`inline:accept:${suggestion.id}`);
-    try {
-      await acceptInlineSuggestion({
-        ctx,
-        suggestion,
-        threadId: thread.id,
-      });
-    } catch {
-      toast.error("Could not apply this inline suggestion.");
-    } finally {
-      setBusyKey(null);
-    }
-  };
-
-  const handleInlineDismiss = async (suggestion: InlineSuggestion) => {
-    setBusyKey(`inline:dismiss:${suggestion.id}`);
-    try {
-      await dismissInlineSuggestion({
-        ctx,
-        suggestion,
-        threadId: thread.id,
-      });
-    } catch {
-      toast.error("Could not dismiss this inline suggestion.");
-    } finally {
-      setBusyKey(null);
-    }
-  };
-
-  const readTime = read.createdAt ? (
-    <ActionRow.Meta>
-      {formatRelativeTime(new Date(read.createdAt))}
-    </ActionRow.Meta>
-  ) : null;
-
-  const summary = (
-    <RichMarkdown
-      content={read.summary}
-      preset="inline"
-      className={
-        read.recommendation
-          ? "text-foreground-secondary"
-          : "text-foreground-primary"
-      }
-    />
-  );
-
-  const recommendation = read.recommendation ? (
-    <RichMarkdown
-      content={read.recommendation}
-      preset="inline"
-      className="text-foreground-primary"
-    />
-  ) : null;
-
-  const inlineSuggestionsRow =
-    inlineSuggestions.length > 0 ? (
-      <InlineSuggestionsRow
-        suggestions={inlineSuggestions}
-        organizationId={ctx.organizationId}
-        busy={busyKey !== null}
-        onAccept={handleInlineAccept}
-        onDismiss={handleInlineDismiss}
-      />
-    ) : null;
-
-  const headerActions = (
-    <>
-      <AgentReadReasoningTrigger reasoning={read.reasoning} />
-      <ActionRow.Dismiss
-        onClick={onClose ?? handleDismissRead}
-        label={onClose ? "Close" : "Dismiss read"}
-      />
-    </>
-  );
-
-  return (
-    <ActionRow.Root tier={urgencyTierFromScore(read.urgencyScore)}>
-      <ActionRow.Header>
-        {variant === "feed" ? (
-          <>
-            <div className="flex flex-col">
-              <div className="flex min-h-8 w-full items-center gap-2 pr-16 text-sm text-foreground-primary">
-                <ThreadRef thread={thread} />
-                {readTime}
-              </div>
-              <div className={cn(treeRowClassName, "items-start")}>
-                {read.recommendation ? (
-                  <TreeSkip
-                    stretchStart={TREE_ROW_GAP_PX}
-                    stretchEnd={TREE_ROW_GAP_PX}
-                  />
-                ) : (
-                  <TreeJoin
-                    isLast
-                    stretchStart={TREE_ROW_GAP_PX}
-                    className="h-8 self-start"
-                  />
-                )}
-                <div className={cn(treeContentClassName, "items-start")}>
-                  {summary}
-                </div>
-              </div>
-              {recommendation ? (
-                <div className={cn(treeRowClassName, "items-start")}>
-                  <TreeJoin
-                    isLast
-                    stretchStart={TREE_ROW_GAP_PX}
-                    className="h-8 self-start"
-                  />
-                  <div className={cn(treeContentClassName, "items-start")}>
-                    {recommendation}
-                  </div>
-                </div>
-              ) : null}
-              {inlineSuggestionsRow ? (
-                <div className="py-1 pl-5">{inlineSuggestionsRow}</div>
-              ) : null}
-            </div>
-            <ActionRow.TopActions>{headerActions}</ActionRow.TopActions>
-          </>
-        ) : (
-          <div className="flex items-start gap-3">
-            <div className="flex min-w-0 flex-1 flex-col gap-1">
-              {summary}
-              {recommendation}
-              {inlineSuggestionsRow ? (
-                <div className="py-1">{inlineSuggestionsRow}</div>
-              ) : null}
-            </div>
-            <div className="flex shrink-0 items-center gap-0.5">
-              {headerActions}
-            </div>
-          </div>
-        )}
-      </ActionRow.Header>
-      <SignalReplyDraftEditor
-        open={replyEditorOpen}
-        draft={replyDraft}
-        contentKey={`${readFingerprint}:${replyEditorRevision}`}
-        onDraftChange={setReplyDraft}
-      />
-      <ActionRow.Actions>
-        {variant === "thread" && readTime ? (
-          <div className="mr-auto">{readTime}</div>
-        ) : null}
-        {variant === "thread" ? (
-          <ActionButton
-            size="sm"
-            variant="ghost"
-            onClick={handleDismissRead}
-            disabled={busyKey !== null}
-          >
-            Dismiss
-          </ActionButton>
-        ) : null}
-        {!replyEditorOpen &&
-          visibleAlternatives.map(({ alternative, index }) => (
-            <ActionButton
-              key={`${thread.id}:alternative:${alternative.kind}:${index}`}
-              size="sm"
-              variant="secondary"
-              onClick={() => handleAlternativeClick(index)}
-              disabled={busyKey !== null}
-            >
-              {ACTION_KIND_LABEL[alternative.kind]}
-            </ActionButton>
-          ))}
-        {replyEditorOpen && (
-          <ActionButton
-            size="sm"
-            variant="ghost"
-            onClick={handleCancelReplyEditor}
-            disabled={busyKey !== null}
-          >
-            Cancel
-          </ActionButton>
-        )}
-        {editingAlternativeReply ? (
-          <ActionButton
-            size="sm"
-            variant="primary"
-            onClick={handleSendReply}
-            disabled={busyKey !== null || replyDraft.trim().length === 0}
-          >
-            Just reply
-          </ActionButton>
-        ) : (
-          read.primary.length > 0 && (
-            <CompoundActionButton
-              label={compoundButtonLabel(orderedSelected, replyEditorOpen)}
-              disabled={busyKey !== null}
-              executeDisabled={
-                orderedSelected.length === 0 ||
-                (replyEditorOpen && replyDraft.trim().length === 0)
-              }
-              onClick={replyEditorOpen ? handleSendReply : handlePrimaryClick}
-              actions={isCompound ? read.primary : null}
-              selectedIndices={selectedIndices}
-              lockedIndex={editingPrimaryReply ? replyIndex : -1}
-              disableToggles={busyKey !== null}
-              onToggle={toggleAction}
-            />
-          )
-        )}
-      </ActionRow.Actions>
-    </ActionRow.Root>
-  );
-}
